Add ConfigureMvcOptions
